+++ lib/nordic_dfu_web.dart
@@ -1,2 +1,5 @@
+export 'package:nordic_dfu_web/models/request_builder_filter.dart';
 export 'package:nordic_dfu_web/src/nordic_web_unsupported.dart'
     if (dart.library.html) 'package:nordic_dfu_web/src/nordic_web.dart';
+
+