+++ lib/src/nordic_web_unsupported.dart
@@ -1,12 +1,19 @@
 import 'dart:async';
 import 'dart:typed_data';
+import '../models/request_builder_filter.dart';
 
 class NordicDfuWeb {
-  static Future startDfu(
-      {required Uint8List uint8list,
-      int dfuDelay = 15,
-      onProgress,
-      onComplete,
-      onError,
-      onLogs}) async {}
+  static Future startDfu({
+    required Uint8List uint8list,
+    int dfuDelay = 15,
+    onProgress,
+    onComplete,
+    onError,
+    onLogs,
+    serviceUuid,
+    controlUuid,
+    packetUuid,
+    buttonUuid,
+    List<RequestBuilderFilter>? requestBuilderFilters,
+  }) async {}
 }